+++ 1.Modulo/basico.js
@@ -1,0 +1,10 @@
+import { Client, GatewayIntentBits } from "discord.js";
+
+export const client = new Client({
+    intents: [GatewayIntentBits.Guilds]
+});
+
+client.once('ready', () => {
+    console.log("Bot listo", client.user.tag);
+});
+
